+++ apps/course-builder-web/src/trpc/api/routers/articles.ts
@@ -63,97 +63,5 @@ export const articlesRouter = createTRPCRouter({
         title: z.string(),
       }),
     )
-    .mutation(async ({ ctx, input }) => {
-      const session = await getServerAuthSession()
-      const user = session?.user
-      const ability = getAbility({ user })
-      if (!user || !ability.can('create', 'Content')) {
-        throw new Error('Unauthorized')
-      }
-
-      const newArticleId = v4()
-
-      await sanityMutation([
-        {
-          createOrReplace: {
-            _id: newArticleId,
-            _type: 'article',
-            state: 'draft',
-            visibility: 'unlisted',
-            title: input.title,
-            slug: {
-              current: slugify(`${input.title}~${nanoid()}`),
-            },
-          },
-        },
-      ])
-
-      const article = await getArticle(newArticleId)
-
-      if (article && session?.user) {
-        console.log('inserting article', { article })
-        await db
-          .insert(contentResource)
-          .values(convertToMigratedArticleResource({ article, ownerUserId: session.user.id }))
-      }
-
-      revalidateTag('articles')
-
-      return article
-    }),
-  update: protectedProcedure.input(ArticleSchema).mutation(async ({ ctx, input }) => {
-    const session = await getServerAuthSession()
-    const user = session?.user
-    const ability = getAbility({ user })
-    if (!user || !ability.can('update', 'Content')) {
-      throw new Error('Unauthorized')
-    }
-
-    const currentArticle = await getArticle(input._id)
-
-    if (input.title !== currentArticle?.title) {
-      const splitSlug = currentArticle?.slug.split('~') || ['', nanoid()]
-      await sanityMutation([
-        {
-          patch: {
-            id: input._id,
-            set: {
-              'slug.current': `${slugify(input.title)}~${splitSlug[1]}`,
-              title: input.title,
-            },
-          },
-        },
-      ])
-    }
-
-    await sanityMutation(
-      [
-        {
-          patch: {
-            id: input._id,
-            set: {
-              ...input,
-              slug: {
-                _type: 'slug',
-                current: input.slug,
-              },
-            },
-          },
-        },
-      ],
-      { returnDocuments: true },
-    ).then((res) => res.results[0].document)
-
-    revalidateTag('articles')
-
-    const updatedArticle = await getArticle(input._id)
-
-    if (updatedArticle) {
-      const dbArticle = convertToMigratedArticleResource({ article: updatedArticle, ownerUserId: user.id })
-      console.log('updating article', { dbArticle })
-      await db.update(contentResource).set(dbArticle).where(eq(contentResource.id, updatedArticle._id))
-    }
-
-    return updatedArticle
-  }),
+    .mutation(async ({ ctx, input }) => {}),
 })
