+++ apps/course-builder-web/src/trpc/api/routers/tips.ts
@@ -1,4 +1,4 @@
-import { revalidateTag } from 'next/cache'
+import { revalidatePath, revalidateTag } from 'next/cache'
 import { TIP_CHAT_EVENT } from '@/inngest/events'
 import { inngest } from '@/inngest/inngest.server'
 import { getAbility } from '@/lib/ability'
@@ -9,6 +9,7 @@ import { contentResource } from '@/server/db/schema'
 import { sanityMutation } from '@/server/sanity.server'
 import { createTRPCRouter, protectedProcedure, publicProcedure } from '@/trpc/api/trpc'
 import { toChicagoTitleCase } from '@/utils/chicagor-title'
+import { guid } from '@/utils/guid'
 import slugify from '@sindresorhus/slugify'
 import { eq } from 'drizzle-orm'
 import { customAlphabet } from 'nanoid'
@@ -54,119 +55,5 @@ export const tipsRouter = createTRPCRouter({
       })
 
       return await getTip(input.tipId)
-    }),
-  create: protectedProcedure
-    .input(
-      z.object({
-        videoResourceId: z.string(),
-        title: z.string(),
-      }),
-    )
-    .mutation(async ({ ctx, input }) => {
-      const session = await getServerAuthSession()
-      const ability = getAbility({ user: session?.user })
-      const user = session?.user
-
-      if (!user || !ability.can('create', 'Content')) {
-        throw new Error('Unauthorized')
-      }
-
-      const newTipId = v4()
-
-      await sanityMutation([
-        {
-          createOrReplace: {
-            _id: newTipId,
-            _type: 'tip',
-            title: input.title,
-            state: 'draft',
-            visibility: 'unlisted',
-            slug: {
-              current: slugify(`${input.title}~${nanoid()}`),
-            },
-            resources: [
-              {
-                _key: v4(),
-                _type: 'reference',
-                _ref: input.videoResourceId,
-              },
-            ],
-          },
-        },
-      ])
-
-      const tip = await getTip(newTipId)
-
-      if (tip) {
-        await db.insert(contentResource).values(convertToMigratedTipResource({ tip, ownerUserId: user.id }))
-      }
-
-      revalidateTag('tips')
-
-      return tip
-    }),
-  update: protectedProcedure
-    .input(
-      z.object({
-        tipId: z.string(),
-        title: z.string(),
-        body: z.string().optional().nullable(),
-      }),
-    )
-    .mutation(async ({ ctx, input }) => {
-      const session = await getServerAuthSession()
-      const user = session?.user
-      const ability = getAbility({ user })
-      if (!user || !ability.can('update', 'Content')) {
-        throw new Error('Unauthorized')
-      }
-
-      const currentTip = await getTip(input.tipId)
-
-      if (!currentTip) {
-        throw new Error('Not found')
-      }
-
-      if (input.title !== currentTip.title) {
-        await sanityMutation([
-          {
-            patch: {
-              id: input.tipId,
-              set: {
-                'slug.current': `${slugify(input.title)}~${nanoid()}`,
-                title: input.title,
-              },
-            },
-          },
-        ])
-      }
-
-      await sanityMutation(
-        [
-          {
-            patch: {
-              id: input.tipId,
-              set: {
-                body: input.body,
-              },
-            },
-          },
-        ],
-        { returnDocuments: true },
-      )
-
-      revalidateTag('tips')
-      revalidateTag(input.tipId)
-
-      const updatedTip = await getTip(input.tipId)
-
-      if (updatedTip) {
-        await db
-          .update(contentResource)
-          .set(convertToMigratedTipResource({ tip: updatedTip, ownerUserId: user.id }))
-          .where(eq(contentResource.id, updatedTip._id))
-      }
-
-      return updatedTip
     }),
 })
